cmake: allow callers to avoid linking libatomic
The current logic wasn't always working, for example when the detected
as 'Clang' compiler was actually zig cc, which doesn't support all the
options CMake expects it to. Instead we let callers avoid linking with
libatomic themselves, by marking -latomic with --as-needed, which will
make the linker link with it only if it is actually used.
This way callers with `CC=clang` can now also pass
`CFLAGS=-march=x86-64-v2` themselves to avoid linking with libatomic.